WordNet mappings
This is the first version of the mapping file.
I made some decisions which may not be 100% semantically correct, in
some cases where I did not find a perfect match.
In particular with businesses that involved occupations, I made the
schema.org business a sub class of the WordNet occupation