Use the rdfs:comment property to fetch type comments
It seems likely that, at some point over the past year, the HTML
structure of the type pages changed to use a <div
property="rdfs:comment"> element to embed the comment, which broke the
scraper.

This enables us to greatly simplify the scraping process, and to make it
more consistent with properties via the use of get_inner_html().

Signed-off-by: Dan Scott <dan@coffeecode.net>